SOC-3276 | NPE Unknown error when comment/like a deleted activity.

+++ component/webui/src/main/java/org/exoplatform/social/webui/activity/BaseUIActivity.java
@@ -612,6 +612,10 @@ public class BaseUIActivity extends UIForm {
     @Override
     public void execute(Event<BaseUIActivity> event) throws Exception {
       BaseUIActivity uiActivity = event.getSource();
+      String activityId = uiActivity.getActivity().getId();
+      if (uiActivity.isNoLongerExisting(activityId, event)) {
+        return;
+      }
       uiActivity.refresh();
       uiActivity.setAllLoaded(true);
       event.getRequestContext().addUIComponentToUpdateByAjax(uiActivity);
@@ -622,6 +626,10 @@ public class BaseUIActivity extends UIForm {
     @Override
     public void execute(Event<BaseUIActivity> event) throws Exception {
       BaseUIActivity uiActivity = event.getSource();
+      String activityId = uiActivity.getActivity().getId();
+      if (uiActivity.isNoLongerExisting(activityId, event)) {
+        return;
+      }
       uiActivity.refresh();
       WebuiRequestContext requestContext = event.getRequestContext();
       String isLikedStr = requestContext.getRequestParameter(OBJECTID);
@@ -635,6 +643,10 @@ public class BaseUIActivity extends UIForm {
     @Override
     public void execute(Event<BaseUIActivity> event) throws Exception {
       BaseUIActivity uiActivity = event.getSource();
+      String activityId = uiActivity.getActivity().getId();
+      if (uiActivity.isNoLongerExisting(activityId, event)) {
+        return;
+      }
       uiActivity.refresh();
       String status = event.getRequestContext().getRequestParameter(OBJECTID);
       CommentStatus commentListStatus = null;
@@ -669,6 +681,10 @@ public class BaseUIActivity extends UIForm {
     @Override
     public void execute(Event<BaseUIActivity> event) throws Exception {
       BaseUIActivity uiActivity = event.getSource();
+      String activityId = uiActivity.getActivity().getId();
+      if (uiActivity.isNoLongerExisting(activityId, event)) {
+        return;
+      }
       uiActivity.refresh();
       WebuiRequestContext requestContext = event.getRequestContext();
       UIFormTextAreaInput uiFormComment = uiActivity.getChild(UIFormTextAreaInput.class);
@@ -695,8 +711,11 @@ public class BaseUIActivity extends UIForm {
     @Override
     public void execute(Event<BaseUIActivity> event) throws Exception {
       BaseUIActivity uiActivity = event.getSource();
-      uiActivity.checkExistingOrNot(uiActivity.getActivity().getId(), event);
-      Utils.getActivityManager().deleteActivity(uiActivity.getActivity().getId());
+      String activityId = uiActivity.getActivity().getId();
+      if (uiActivity.isNoLongerExisting(activityId, event)) {
+        return;
+      }
+      Utils.getActivityManager().deleteActivity(activityId);
       UIActivitiesContainer activitiesContainer = uiActivity.getParent();
       activitiesContainer.removeChildById(uiActivity.getId());
       activitiesContainer.removeActivity(uiActivity.getActivity());
@@ -714,6 +733,10 @@ public class BaseUIActivity extends UIForm {
     @Override
     public void execute(Event<BaseUIActivity> event) throws Exception {
       BaseUIActivity uiActivity = event.getSource();
+      String activityId = uiActivity.getActivity().getId();
+      if (uiActivity.isNoLongerExisting(activityId, event)) {
+        return;
+      }
       WebuiRequestContext requestContext = event.getRequestContext();
       Utils.getActivityManager().deleteComment(uiActivity.getActivity().getId(),
                                                requestContext.getRequestParameter(OBJECTID));
@@ -732,17 +755,18 @@ public class BaseUIActivity extends UIForm {
     }
   }
 
-  protected void checkExistingOrNot(String activityId, Event<BaseUIActivity> event) {
+  protected boolean isNoLongerExisting(String activityId, Event<BaseUIActivity> event) {
     ExoSocialActivity existingActivity = Utils.getActivityManager().getActivity(activityId);
     if (existingActivity == null) {
       UIApplication uiApplication = event.getRequestContext().getUIApplication();
-      uiApplication.addMessage(new ApplicationMessage("UIComposer.msg.info.Activity_Comment_Deleted",
+      uiApplication.addMessage(new ApplicationMessage("BaseUIActivity.msg.info.Activity_No_Longer_Exist",
                                                     null,
                                                     ApplicationMessage.INFO));
-      return;
+      return true;
     }
+    return false;
   }
-
+  
   /**
    * Allow child can be override this method to process I18N
    * @param activity
